moved ai part to entityparts
