+++ CommonMap/src/main/java/dk/sdu/mmmi/modulemon/CommonMap/Data/EntityParts/AIControlPart.java
@@ -2,9 +2,11 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package dk.sdu.mmmi.modulemon.CommonMap.Data;
+package dk.sdu.mmmi.modulemon.CommonMap.Data.EntityParts;
 
+import dk.sdu.mmmi.modulemon.CommonMap.Data.Entity;
 import dk.sdu.mmmi.modulemon.CommonMap.Data.EntityParts.EntityPart;
+import dk.sdu.mmmi.modulemon.CommonMap.Data.World;
 import dk.sdu.mmmi.modulemon.common.data.GameData;
 
 /**
