+++ NPC/src/main/java/dk/sdu/mmmi/modulemon/NPC/NPCControlSystem.java
@@ -5,12 +5,8 @@
 package dk.sdu.mmmi.modulemon.NPC;
 
 import com.badlogic.gdx.graphics.Texture;
-import dk.sdu.mmmi.modulemon.CommonMap.Data.*;
 import dk.sdu.mmmi.modulemon.CommonMap.Data.Entity;
-import dk.sdu.mmmi.modulemon.CommonMap.Data.EntityParts.InteractPart;
-import dk.sdu.mmmi.modulemon.CommonMap.Data.EntityParts.MovingPart;
-import dk.sdu.mmmi.modulemon.CommonMap.Data.EntityParts.PositionPart;
-import dk.sdu.mmmi.modulemon.CommonMap.Data.EntityParts.SpritePart;
+import dk.sdu.mmmi.modulemon.CommonMap.Data.EntityParts.*;
 import dk.sdu.mmmi.modulemon.common.data.GameData;
 import dk.sdu.mmmi.modulemon.CommonMap.Data.World;
 import dk.sdu.mmmi.modulemon.CommonMap.Services.IEntityProcessingService;
